feat: added logic to increment the count of the current date
